Make changes with help of solution
Made a few changes using the solution css file as help.
A very important thing to note is that the height
 wasn't necessary for the cards.
The heigh is automatic, if I remember correctly,
 so only with the width it looks incredibly better.

Another very important thing was erasing the flex
 property from the sidebar.
A flex-shrink: 0 was enough to prevent it from shrinking.
The value for basis is unnecesarry since the width was already declared.

And witout the flex-grow value declaration
 it still fills the space correctly.
The margin and the padding of the sidebar
 links were tweaked a litte bit, too.
And the list items were given a margin-bottom of 16px.